Only try to remove base conf if it exists
This won't necessarily be present, and when it's not (e.g. trying to build just the merlin model pack and no others),
it previously threw a FileNotFoundError

Check if it exists instead of using ignore_errors param of shutil.rmtree as we don't want to ignore
e.g. permissions errors that actually are a problem.

+++ kazu/utils/build_and_test_model_packs.py
@@ -182,8 +182,9 @@ class ModelPackBuilder:
         else:
             # the base conf may have been copied over if the base model pack was used,
             # so we get rid of it here if required
-            print(f"removing base configuration for {uncached_model_pack_path.name}")
-            shutil.rmtree(conf_path)
+            if conf_path.exists():
+                print(f"removing base configuration for {uncached_model_pack_path.name}")
+                shutil.rmtree(conf_path)
 
         print(f"preparing to build model pack {model_pack_build_path.name}")
         copy_tree(str(uncached_model_pack_path), str(model_pack_build_path))
